Added access control.

+++ app/Http/Controllers/PostsController.php
@@ -8,11 +8,23 @@ use DB;
 
 class PostsController extends Controller
 {
+    
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
+
     public function index()
     {
         // Get all post
@@ -88,6 +100,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unathorized Page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -124,6 +142,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unathorized Page');
+        }
+        
         $post->delete();
 
         // Redirect to post with success message.
